Split SQG test suite matching new separation in the code under test
